requirements.txt
# Core LangChain ecosystem
langchain
langchain-community
langchain-experimental

# LLM providers
langchain-openai
langchain-google-genai
`
# Vector DB
langchain-chroma

# Document loaders
pypdf
beautifulsoup4

# Environment
python-dotenv

# UI
streamlit

# Utilities
email-validator
